Extract `BuildResult` into `harmonia-store-build-result` crate
Same pattern as `ValidPathInfo` / `harmonia-store-path-info`: pure types
and JSON serialization live in the new crate, wire protocol
`NixDeserialize`/`NixSerialize` impls remain in `harmonia-protocol`. The
motivation is the same too --- both `BuildResult` and `ValidPathInfo` are a bit more
implementation-opinated / not the only way of doing things, so I don't
want them to live in `harmonia-store-core`.

Type changes for protocol independence:
- `error_msg`: `DaemonString` (`Bytes`) → `Vec<u8>`
- `times_built`: `DaemonInt` (`c_uint`) → `u32`
- `start_time`/`stop_time`: `DaemonTime` (`time_t`) → `i64`
- `Microseconds` moved from `types2.rs` to the new crate
