+++ harmonia-protocol/Cargo.toml
@@ -33,26 +33,34 @@ tokio-util       = { workspace = true, features = [ "io" ] }
 tracing          = { workspace = true }
 
 # Internal dependencies
-harmonia-nar             = { version = "3.0.0" }
-harmonia-protocol-derive = { version = "3.1.0" }
-harmonia-store-core      = { version = "0.0.0-alpha.0" }
-harmonia-store-path-info = { path = "../harmonia-store-path-info" }
-harmonia-utils-hash      = { version = "0.0.0-alpha.0" }
-harmonia-utils-io        = { version = "0.0.0-alpha.0" }
-harmonia-utils-test      = { version = "0.0.0-alpha.0", optional = true }
+harmonia-nar                = { version = "3.0.0" }
+harmonia-protocol-derive    = { version = "3.1.0" }
+harmonia-store-build-result = { path = "../harmonia-store-build-result" }
+harmonia-store-core         = { version = "0.0.0-alpha.0" }
+harmonia-store-path-info    = { path = "../harmonia-store-path-info" }
+harmonia-utils-hash         = { version = "0.0.0-alpha.0" }
+harmonia-utils-io           = { version = "0.0.0-alpha.0" }
+harmonia-utils-test         = { version = "0.0.0-alpha.0", optional = true }
 
 [dev-dependencies]
-harmonia-nar        = { version = "3.0.0", features = [ "test" ] }
-harmonia-store-core = { version = "0.0.0-alpha.0", features = [ "test" ] }
-harmonia-utils-test = { version = "0.0.0-alpha.0" }
-hex-literal         = { workspace = true }
-proptest            = { workspace = true }
-proptest-derive     = { workspace = true }
-rstest              = { workspace = true }
-tokio-test          = { workspace = true }
+harmonia-nar                = { version = "3.0.0", features = [ "test" ] }
+harmonia-store-build-result = { path = "../harmonia-store-build-result", features = [ "test" ] }
+harmonia-store-core         = { version = "0.0.0-alpha.0", features = [ "test" ] }
+harmonia-utils-test         = { version = "0.0.0-alpha.0" }
+hex-literal                 = { workspace = true }
+proptest                    = { workspace = true }
+proptest-derive             = { workspace = true }
+rstest                      = { workspace = true }
+tokio-test                  = { workspace = true }
 
 [features]
-test = [ "proptest", "proptest-derive", "harmonia-utils-test", "harmonia-utils-hash/test" ]
+test = [
+    "proptest",
+    "proptest-derive",
+    "harmonia-utils-test",
+    "harmonia-utils-hash/test",
+    "harmonia-store-build-result/test",
+]
 
 [[test]]
 name = "json_upstream"
