+++ harmonia-protocol/src/lib.rs
@@ -27,7 +27,7 @@
 //! 3. **Well-specified**: Document wire format
 //! 4. **Type-safe**: Use strong types for protocol messages
 
-pub mod build_result;
+mod build_result;
 pub mod daemon_wire;
 pub mod de;
 pub mod log;
