+++ harmonia-store-build-result/Cargo.toml
@@ -1,0 +1,36 @@
+# SPDX-FileCopyrightText: 2025 Obsidian Systems
+# SPDX-License-Identifier: MIT
+
+[package]
+name                 = "harmonia-store-build-result"
+version.workspace    = true
+authors.workspace    = true
+edition.workspace    = true
+license.workspace    = true
+homepage.workspace   = true
+repository.workspace = true
+description          = "BuildResult types and JSON serialization for Nix build results"
+readme               = "README.md"
+
+[dependencies]
+harmonia-store-core = { path = "../harmonia-store-core" }
+num_enum            = { workspace = true }
+proptest            = { workspace = true, optional = true }
+proptest-derive     = { workspace = true, optional = true }
+serde               = { workspace = true }
+serde_json          = { workspace = true }
+
+[dev-dependencies]
+harmonia-store-build-result = { path = ".", features = [ "test" ] }
+harmonia-utils-test         = { version = "0.0.0-alpha.0" }
+serde_json                  = { workspace = true }
+
+[features]
+test = [ "proptest", "proptest-derive", "harmonia-store-core/test" ]
+
+[[test]]
+name = "json_upstream"
+path = "tests/json_upstream/mod.rs"
+
+[lints]
+workspace = true
